+++ lib/src/main/java/org/asamk/signal/manager/helper/AttachmentHelper.java
@@ -1,6 +1,7 @@
 package org.asamk.signal.manager.helper;
 
 import org.asamk.signal.manager.api.AttachmentInvalidException;
+import org.asamk.signal.manager.api.Message.AttachmentDimensions;
 import org.asamk.signal.manager.config.ServiceConfig;
 import org.asamk.signal.manager.internal.SignalDependencies;
 import org.asamk.signal.manager.storage.AttachmentStore;
@@ -52,9 +53,10 @@ public class AttachmentHelper {
 
     public List<SignalServiceAttachment> uploadAttachments(
             final List<String> attachments,
+            final List<AttachmentDimensions> dimensions,
             boolean voiceNote
     ) throws AttachmentInvalidException, IOException {
-        final var attachmentStreams = createAttachmentStreams(attachments, voiceNote);
+        final var attachmentStreams = createAttachmentStreams(attachments, dimensions, voiceNote);
 
         try {
             // Upload attachments here, so we only upload once even for multiple recipients
@@ -70,27 +72,25 @@ public class AttachmentHelper {
         }
     }
 
-    public List<SignalServiceAttachment> uploadAttachments(final List<String> attachments) throws AttachmentInvalidException, IOException {
-        return uploadAttachments(attachments, false);
-    }
-
     private List<SignalServiceAttachmentStream> createAttachmentStreams(
             List<String> attachments,
+            List<AttachmentDimensions> dimensions,
             boolean voiceNote
     ) throws AttachmentInvalidException, IOException {
         if (attachments == null) {
             return null;
         }
         final var signalServiceAttachments = new ArrayList<SignalServiceAttachmentStream>(attachments.size());
-        for (var attachment : attachments) {
-            final var attachmentStream = getAttachmentStream(attachment, voiceNote);
-            signalServiceAttachments.add(attachmentStream);
+        for (var i = 0; i < attachments.size(); i++) {
+            final var size = i < dimensions.size() ? dimensions.get(i) : AttachmentDimensions.UNKNOWN;
+            signalServiceAttachments.add(getAttachmentStream(attachments.get(i), size, voiceNote));
         }
         return signalServiceAttachments;
     }
 
     private SignalServiceAttachmentStream getAttachmentStream(
             final String attachment,
+            final AttachmentDimensions dimensions,
             final boolean voiceNote
     ) throws AttachmentInvalidException {
         try {
@@ -116,6 +116,7 @@ public class AttachmentHelper {
             return AttachmentUtils.createAttachmentStream(streamDetails,
                     streamDetailsAndFileName.second(),
                     voiceNote,
+                    dimensions,
                     uploadSpec);
         } catch (IOException e) {
             throw new AttachmentInvalidException(attachment, e);
@@ -130,7 +131,7 @@ public class AttachmentHelper {
     }
 
     public SignalServiceAttachmentPointer uploadAttachment(String attachment) throws IOException, AttachmentInvalidException {
-        final var attachmentStream = getAttachmentStream(attachment, false);
+        final var attachmentStream = getAttachmentStream(attachment, AttachmentDimensions.UNKNOWN, false);
         return uploadAttachment(attachmentStream);
     }