+++ src/main/java/org/asamk/signal/dbus/DbusSignalImpl.java
@@ -242,6 +242,7 @@ public class DbusSignalImpl implements Signal, AutoCloseable {
         try {
             final var message = new Message(messageText,
                     attachments,
+                    List.of(),
                     false,
                     false,
                     List.of(),
@@ -409,6 +410,7 @@ public class DbusSignalImpl implements Signal, AutoCloseable {
         try {
             final var message = new Message(messageText,
                     attachments,
+                    List.of(),
                     false,
                     false,
                     List.of(),
@@ -456,6 +458,7 @@ public class DbusSignalImpl implements Signal, AutoCloseable {
         try {
             final var message = new Message(messageText,
                     attachments,
+                    List.of(),
                     false,
                     false,
                     List.of(),